requirements.txt
openai
streamlit
requests
pydub
tqdm
yt_dlp
pytube
python-dotenv==0.19.1
imageio-ffmpeg